update: move tree-related fns to ecs helpers
