+++ packages/@dcl/ecs/src/index.ts
@@ -3,6 +3,7 @@ export * from './engine'
 export * from './schemas'
 export * from './runtime/initialization'
 export * from './runtime/types'
+export * from './runtime/helpers'
 
 export { cyclicParentingChecker } from './systems/cyclicParentingChecker'
 export * from './systems/events'
