+++ packages/@dcl/ecs/src/runtime/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './tree'
